+++ Food-Express/Food-Express/Screens/001-MainPage/View/DetailsFood/DetailsFoodViewController.swift
@@ -10,7 +10,7 @@ import AlamofireImage
 
 final class DetailsFoodViewController: UIViewController {
     
-    private let viewModel: DetailsFoodModel
+    private let viewModel: FoodViewModel
     
     @IBOutlet private weak var foodNameLbl: UILabel!
     @IBOutlet private weak var foodPriceLbl: UILabel!
@@ -20,7 +20,7 @@ final class DetailsFoodViewController: UIViewController {
 
     private var amount = 1
 
-    init(viewModel: DetailsFoodModel) {
+    init(viewModel: FoodViewModel) {
         self.viewModel = viewModel
         super.init(nibName: "DetailsFoodViewController", bundle: nil)
     }
